+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-All notable changes to this project will be documented in this file.
+All notable changes to Gubug will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
@@ -9,7 +9,7 @@ Tags: Added, Changed, Deprecated, Removed, Fixed, Security
 
 ---
 
-## [Unreleased]
+## [NEXT] - YYYY-MM-DD
 
 ---
 
@@ -19,4 +19,4 @@ Tags: Added, Changed, Deprecated, Removed, Fixed, Security
 * Graduated from Alpha version
 
 ---
-[Unreleased]: https://github.com/qaharmdz/gubug/compare/v1.0.0-beta.1...HEAD
+[NEXT]: https://github.com/qaharmdz/gubug/compare/v1.0.0-beta.1...HEAD